refactor: drop multiselect, use single + non-locational for hits
