+++ src/scripts/ui/select/index.ts
@@ -4,9 +4,6 @@ export type {
   SelectOption,
   SingleSelectOptions,
   SingleSelectResult,
-  MultiSelectOptions,
-  MultiSelectResult,
 } from "./types";
 
 export { createSingleSelect } from "./single";
-export { createMultiSelect } from "./multi";
